refactor(improve): move loop/post-loop/maintenance stages to improve/loop-stages.ts (D1c step C)
Move runImproveLoopStage (619 LOC), runImprovePostLoopStage (167), and
runImproveMaintenancePasses (446) out of improve.ts into improve/loop-stages.ts.
improve.ts 2,693 -> 1,443 LOC (from 5,395 at the start of D1). The 3 stages have
zero real local-symbol deps on improve.ts; one-directional value import +
type-only back-import of the shared result/context types (no runtime cycle).
runImproveMaintenancePasses re-exported for the db-locking test. Verbatim move;
lint 0 / tsc 0 / unit 5274-0.
